Refactor file paths
